+++ MeoMeoApplication/Services/ProductDetailService.cs
@@ -15,36 +15,57 @@ namespace MeoMeo.Application.Services
     {
         private readonly IProductsDetailRepository _repository;
         private readonly IMapper _mapper;
-    }
-    public async Task<Product> CreateProductDetailAsync(CreateOrUpdateProductDTO product)
-    {
-        var mappedProduct = _mapper.Map<Product>(product);
-        mappedProduct.Id = Guid.NewGuid();
-        return await _repository.AddAsync(mappedProduct);
-    }
-    public async Task<IEnumerable<ProductDetail>> GetAllAsync()
-    {
-        return await _repository.GetAllAsync();
-    }
 
-    public async Task<ProductDetail?> GetByIdAsync(Guid id)
-    {
-        return await _repository.GetByIdAsync(id);
-    }
+        public ProductDetailService(IProductsDetailRepository repository, IMapper mapper)
+        {
+            _repository = repository;
+            _mapper = mapper;
+        }
 
-    public async Task AddAsync(ProductDetail entity)
-    {
-        await _repository.AddAsync(entity);
-    }
+        public async Task<ProductDetail> AddProductDetailAsync(CreateOrUpdateProductDetailDTO dto)
+        {
+            var entity = _mapper.Map<ProductDetail>(dto);
+            entity.Id = Guid.NewGuid();
+            return await _repository.AddProductAsync(entity);
+        }
 
-    public async Task UpdateAsync(ProductDetail entity)
-    {
-        await _repository.UpdateAsync(entity);
-    }
+        public async Task<ProductDetail> DeleteProductDetailAsync(Guid id)
+        {
+            var existing = await _repository.GetByProductIdAsync(id);
+            if (existing == null)
+                throw new Exception("Sản phẩm không tồn tại.");
 
-    public async Task DeleteAsync(Guid id)
-    {
-        await _repository.DeleteAsync(id);
+            await _repository.DeleteAsync(id);
+
+            return existing; 
+        }
+
+        public async Task<IEnumerable<ProductDetail>> GetProductDetailAllAsync()
+        {
+            return await _repository.GetAllAsync();
+        }
+
+        public async Task<ProductDetail> GetProductDetailByIdAsync(Guid id)
+        {
+          return await _repository.GetByProductIdAsync(id);
+        }
+
+        public async Task<ProductDetail> UpdateProductDetailAsync(CreateOrUpdateProductDetailDTO entity)
+        {
+            var existing = await _repository.GetByProductIdAsync(entity.Id);
+            if (existing == null)
+                throw new Exception("Sản phẩm không tồn tại.");
+
+           
+            _mapper.Map(entity, existing);
+
+            await _repository.UpdateProductAsync(existing);
+
+            return existing;
+        }
     }
 }
-}
+
+
+
+
